refactor(eslint): extract rule sets
For improved maintainability and seperation of concerns.
